Fixed a lot of balancing and progression issues.
Changed a lot of enemies's variables to make the progression more fun for the player and last longer. Fixed the recovery time on the enemies. Adjusted the ragdoll breaking (Still has to be worked on). Balls now spawn less frequently and more spread out. Fixed issue of restarting the scene straight away after game over.